requirements.txt
gunicorn
flask
tensorflow==2.11
pillow
numpy
protobuf
keras